test: new render config
