+++ packages/integrations/markdoc/test/fixtures/content-collections/src/content/blog/post-1.mdoc
@@ -1,0 +1,7 @@
+---
+title: Post 1
+---
+
+## Post 1
+
+This is the contents of post 1.
