+++ packages/integrations/markdoc/test/fixtures/content-collections/src/content/blog/post-2.mdoc
@@ -1,0 +1,7 @@
+---
+title: Post 2
+---
+
+## Post 2
+
+This is the contents of post 2.
