+++ packages/integrations/markdoc/test/fixtures/content-collections/src/content/blog/post-3.mdoc
@@ -1,0 +1,7 @@
+---
+title: Post 3
+---
+
+## Post 3
+
+This is the contents of post 3.
