+++ packages/integrations/markdoc/test/fixtures/content-collections/src/pages/entry.json.js
@@ -3,7 +3,7 @@ import { stringify } from 'devalue';
 import { stripRenderFn } from '../../utils.js';
 
 export async function get() {
-	const post = await getEntryBySlug('blog', 'simple');
+	const post = await getEntryBySlug('blog', 'post-1');
 	return {
 		body: stringify(stripRenderFn(post)),
 	};
