+++ packages/integrations/markdoc/test/fixtures/render-simple/src/content/blog/simple.mdoc
