+++ packages/integrations/markdoc/test/fixtures/render-simple/src/pages/index.astro
@@ -1,5 +1,6 @@
 ---
 import { getEntryBySlug } from "astro:content";
+
 const post = await getEntryBySlug('blog', 'simple');
 const { Content } = await post.render();
 ---
@@ -10,7 +11,7 @@ const { Content } = await post.render();
 	<meta charset="UTF-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Content - Simple</title>
+	<title>Content</title>
 </head>
 <body>
 	<Content />
