+++ packages/integrations/markdoc/test/fixtures/render-with-components/src/components/Code.astro
