+++ packages/integrations/markdoc/test/fixtures/render-with-components/src/components/CustomMarquee.astro
