+++ packages/integrations/markdoc/test/fixtures/render-with-components/src/content/blog/with-components.mdoc
