+++ packages/integrations/markdoc/test/fixtures/render-with-config/markdoc.config.mjs
@@ -1,0 +1,13 @@
+export default {
+	variables: {
+		countries: ['ES', 'JP'],
+	},
+	functions: {
+		includes: {
+			transform(parameters) {
+				const [array, value] = Object.values(parameters);
+				return Array.isArray(array) ? array.includes(value) : false;
+			},
+		},
+	},
+}
