+++ packages/integrations/markdoc/test/fixtures/render-with-config/src/content/blog/with-config.mdoc
