+++ packages/integrations/markdoc/test/fixtures/render-with-config/src/pages/index.astro
@@ -11,9 +11,9 @@ const { Content } = await post.render();
 	<meta charset="UTF-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Content - with config</title>
+	<title>Content</title>
 </head>
 <body>
-	<Content />
+	<Content />	
 </body>
 </html>
